@@ -27,7 +27,7 @@
 		<url/>
 	</scm>
 	<properties>
-		<java.version>17</java.version>
+		<java.version>21</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -100,8 +100,14 @@
 			    <groupId>org.springframework.boot</groupId>
 			    <artifactId>spring-boot-starter-security</artifactId>
 			</dependency>
-			
-			<!-- JWT Dependencies -->
+			<!-- PDF Generation with iText -->
+			<dependency>
+			    <groupId>com.itextpdf</groupId>
+			    <artifactId>itext7-core</artifactId>
+			    <version>8.0.3</version>
+			    <type>pom</type>
+			</dependency>
+			<!-- JWT -->
 			<dependency>
 			    <groupId>io.jsonwebtoken</groupId>
 			    <artifactId>jjwt-api</artifactId>
@@ -148,7 +154,26 @@
 					</excludes>
 				</configuration>
 			</plugin>
-			
+			<plugin>
+            <groupId>org.jacoco</groupId>
+            <artifactId>jacoco-maven-plugin</artifactId>
+            <version>0.8.12</version>
+            <executions>
+                <execution>
+                    <id>prepare-agent</id>
+                    <goals>
+                        <goal>prepare-agent</goal>
+                    </goals>
+                </execution>
+                <execution>
+                    <id>report</id>
+                    <phase>test</phase>
+                    <goals>
+                        <goal>report</goal>
+                    </goals>
+                </execution>
+            </executions>
+        </plugin>		
 			
 		</plugins>
 	</build>
